Move annotation to tck
